requirements.txt
asgiref==3.2.3
astroid==2.3.3
beautifulsoup4==4.8.2
bootstrap4==0.1.0
colorama==0.4.3
dj-static==0.0.6
Django==3.0.3
django-autoslug==1.9.6
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-rest-authtoken==1.2.4
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
wrapt==1.11.2
